fix(code-refactoring): move exception mapper to ExceptionHandler
