+++ src/main/kotlin/it/pagopa/wallet/controllers/WalletController.kt
@@ -15,10 +15,8 @@ import kotlinx.coroutines.reactor.mono
 import lombok.extern.slf4j.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
-import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.server.ServerWebExchange
 import org.springframework.web.util.UriComponentsBuilder
@@ -245,28 +243,5 @@ class WalletController(
                 .filter { header: String -> header.startsWith("Bearer ") }
                 .map { header: String -> header.substring("Bearer ".length) }
         )
-    }
-
-    @ExceptionHandler(WalletServiceStatusConflictException::class)
-    fun walletServiceStatusConflictExceptionHandler(
-        exception: WalletServiceStatusConflictException
-    ): ResponseEntity<WalletServicesPartialUpdateDto> {
-        return ResponseEntity.status(HttpStatus.CONFLICT)
-            .body(
-                WalletServicesPartialUpdateDto().apply {
-                    updatedServices =
-                        exception.updatedServices.map {
-                            WalletServiceDto()
-                                .name(ServiceNameDto.valueOf(it.key.name))
-                                .status(WalletServiceStatusDto.valueOf(it.value.name))
-                        }
-                    failedServices =
-                        exception.failedServices.map {
-                            ServiceDto()
-                                .name(ServiceNameDto.valueOf(it.key.name))
-                                .status(ServiceStatusDto.valueOf(it.value.name))
-                        }
-                }
-            )
     }
 }
